[proxy] split code to files
